majit-metainterp: take the vable array index signed and load at the item width
`bhimpl_{get,set}arrayitem_vable_*` (`blackhole.py:1374-1403`) take the index
with argcode `i` and hand it to `bh_{get,set}arrayitem_gc_*` signed.  The six
handlers cast the register to `usize`, so a negative index became a huge
offset that `data_ptr.add` turned into an address far past the array.  They
now pass the `i64` through and `vable_{read,write}_array_item` refuse a
negative one.

A null array field returned `0` from the read path and silently dropped the
write.  Upstream reaches the items through `bh_getfield_gc_r` and then
`bh_{get,set}arrayitem_gc_*`, which faults; both paths now assert instead.

`item_size` was used for the stride only: every non-`Ref` read was
`ptr::read::<i64>` and every non-`Ref` write an 8-byte store.
`bh_getarrayitem_gc_i` dispatches on `(item_size, is_item_signed)`
(`majit-backend` `model.rs`), so `VableArrayInfo` gains `item_signed` from
`arraydescr.is_item_signed()` and the two helpers dispatch on the pair,
panicking on a width with no load or store the way `llmodel.py:478` raises
`NotImplementedError`.

The write path also gains the read path's `jit_strict_mode` bounds assert.

+++ majit/majit-metainterp/src/blackhole.rs
@@ -11643,7 +11643,7 @@ fn handler_getarrayitem_vable_i(
     p: usize,
 ) -> Result<usize, DispatchError> {
     let vable = bh.registers_r[code[p] as usize];
-    let index = bh.registers_i[code[p + 1] as usize] as usize;
+    let index = bh.registers_i[code[p + 1] as usize];
     let vinfo = vable_clear_token_and_get_vinfo(bh, vable);
     let (field_descr, p) = read_descr(bh, code, p + 2);
     let array_idx = field_descr.as_vable_array_index();
@@ -11661,7 +11661,7 @@ fn handler_getarrayitem_vable_r(
 ) -> Result<usize, DispatchError> {
     let nbody_debug = crate::nbody_debug_enabled();
     let vable = bh.registers_r[code[p] as usize];
-    let index = bh.registers_i[code[p + 1] as usize] as usize;
+    let index = bh.registers_i[code[p + 1] as usize];
     let vinfo = vable_clear_token_and_get_vinfo(bh, vable);
     let (field_descr, p) = read_descr(bh, code, p + 2);
     let array_idx = field_descr.as_vable_array_index();
@@ -11684,7 +11684,7 @@ fn handler_setarrayitem_vable_i(
     p: usize,
 ) -> Result<usize, DispatchError> {
     let vable = bh.registers_r[code[p] as usize];
-    let index = bh.registers_i[code[p + 1] as usize] as usize;
+    let index = bh.registers_i[code[p + 1] as usize];
     let value = bh.registers_i[code[p + 2] as usize];
     let vinfo = vable_clear_token_and_get_vinfo(bh, vable);
     let (field_descr, p) = read_descr(bh, code, p + 3);
@@ -11703,7 +11703,7 @@ fn handler_setarrayitem_vable_r(
 ) -> Result<usize, DispatchError> {
     let nbody_debug = crate::nbody_debug_enabled();
     let vable = bh.registers_r[code[p] as usize];
-    let index = bh.registers_i[code[p + 1] as usize] as usize;
+    let index = bh.registers_i[code[p + 1] as usize];
     let value = bh.registers_r[code[p + 2] as usize];
     let vinfo = vable_clear_token_and_get_vinfo(bh, vable);
     let (field_descr, p) = read_descr(bh, code, p + 3);
@@ -12324,7 +12324,7 @@ fn handler_getarrayitem_vable_f(
     p: usize,
 ) -> Result<usize, DispatchError> {
     let vable = bh.registers_r[code[p] as usize];
-    let index = bh.registers_i[code[p + 1] as usize] as usize;
+    let index = bh.registers_i[code[p + 1] as usize];
     let vinfo = vable_clear_token_and_get_vinfo(bh, vable);
     let (field_descr, p) = read_descr(bh, code, p + 2);
     let array_idx = field_descr.as_vable_array_index();
@@ -12342,7 +12342,7 @@ fn handler_setarrayitem_vable_f(
     p: usize,
 ) -> Result<usize, DispatchError> {
     let vable = bh.registers_r[code[p] as usize];
-    let index = bh.registers_i[code[p + 1] as usize] as usize;
+    let index = bh.registers_i[code[p + 1] as usize];
     let value = bh.registers_f[code[p + 2] as usize];
     let vinfo = vable_clear_token_and_get_vinfo(bh, vable);
     let (field_descr, p) = read_descr(bh, code, p + 3);
